feat: Add currency conversion support to building entity
This commit adds support for currency conversion to the building entity. It introduces two new properties, baseCurrency and auxiliaryCurrency, which are related to the currencyList entity. These properties allow for storing the base and auxiliary currencies for a building. This will be useful for future currency conversion calculations.

The changes also include updating the findOneByOwner method in the buildings service to include the baseCurrency and auxiliaryCurrency relations when retrieving a building by owner.

Lastly, the currency-value-per-day controller has been updated to use a more descriptive route for currency conversion, changing the route from '/:currency/:value/:date' to '/convert/:currency/:value/:date' and '/:currency/:value' to '/convert/:currency/:value'.